Commit attempt one.  Please don't explode
Tried hooking up a script that changes the color of the Floor object.
Not working for some random reason, but the code snippet for changing
color (renderer.material.SetColor()) works when dropped into the
PlayerController.  No idea what the deal is.

Added some material properties to the floor as well.  I think I'm going
to continue using the floor object as a way to pick up Unity.